requirements.txt
streamlit
pandasai
python-dotenv
bs4
plotly
pandas
openai
PyYAML